// vietnam/assets/ScriptCore/protocol/player.d.ts
export namespace player {
// ...
    interface IPlayerData {

        /** PlayerData AccId */
        AccId?: (string|null);

        /** PlayerData Platform */
        Platform?: (string|null);

        /** PlayerData Channel */
        Channel?: (string|null);

        /** PlayerData Name */
        Name?: (string|null);

        /** PlayerData SnId */
        SnId?: (number|null);
// ...
    }
// ...
    interface ISCPlayerFlag {

        /** SCPlayerFlag PlayerId */
        PlayerId?: (number|null);

        /** SCPlayerFlag Flag */
        Flag?: (number|null);
    }
// ...
    interface ISCGetSliderVerifyCode {

        /** SCGetSliderVerifyCode OpRetCode */
        OpRetCode?: (player.OpResultCode|null);

        /** SCGetSliderVerifyCode Pos */
        Pos?: (number|null);
    }
// ...
}